feat: add deprecation warning for environment variables
